+++ Sources/SwiftletModel/Relationship/RelationDelete.swift
@@ -152,53 +152,3 @@ public extension EntityModelProtocol {
         )
     }
 }
-//
-//extension EntityModelProtocol {
-//    func unlink<Child, Cardinality, Constraint>(
-//        _ children: [Child.ID],
-//        _ keyPath: KeyPath<Self, OneWayRelation<Child, Cardinality, Constraint>>
-//
-//    ) -> Links<Self, Child> {
-//
-//        Links(
-//            direct: [Link(
-//                parent: id,
-//                children: children,
-//                attribute: LinkAttribute(
-//                    name: keyPath.name,
-//                    updateOption: .remove
-//                )
-//            )],
-//            inverse: []
-//        )
-//    }
-//
-//    func unlink<Child, Cardinality, Constraint, InverseRelation, InverseConstraint>(
-//        _ children: [Child.ID],
-//        _ keyPath: KeyPath<Self, MutualRelation<Child, Cardinality, Constraint>>,
-//        inverse: KeyPath<Child, MutualRelation<Self, InverseRelation, InverseConstraint>>
-//
-//    ) -> Links<Self, Child> {
-//
-//        Links(
-//            direct: [Link(
-//                parent: id,
-//                children: children,
-//                attribute: LinkAttribute(
-//                    name: keyPath.name,
-//                    updateOption: .remove
-//                )
-//            )],
-//            inverse: children.map { child in
-//                Link(
-//                    parent: child,
-//                    children: [id],
-//                    attribute: LinkAttribute(
-//                        name: inverse.name,
-//                        updateOption: .remove
-//                    )
-//                )
-//            }
-//        )
-//    }
-//}
