+++ Sources/SwiftletModel/Relationship/RelationSave.swift
@@ -63,37 +63,6 @@ public extension EntityModelProtocol {
 
 private extension EntityModelProtocol {
 
-//    func saveRelation<Child, Cardinality, Constraint>(
-//        at keyPath: KeyPath<Self, OneWayRelation<Child, Cardinality, Constraint>>,
-//        in context: inout Context) {
-//
-//        context.updateLinks(link(relationIds(keyPath), keyPath))
-//    }
-//
-//    func saveRelation<Child, Cardinality, Constraint, InverseRelation, InverseConstraint>(
-//        at keyPath: KeyPath<Self, MutualRelation<Child, Cardinality, Constraint>>,
-//        inverse: KeyPath<Child, MutualRelation<Self, InverseRelation, InverseConstraint>>,
-//        in context: inout Context) {
-//
-//        let children = relationIds(keyPath)
-//        switch relation(keyPath).directLinkUpdateOption {
-//        case .append:
-//            context.updateLinks(link(children, keyPath, inverse: inverse))
-//        case .replace, .remove:
-//            let enititesToKeep = Set(children)
-//            let oddExisingChildren = context
-//                .getChildren(for: Self.self, relationName: keyPath.name, id: id)
-//                .compactMap { Child.ID($0) }
-//                .filter { !enititesToKeep.contains($0) }
-//
-//            context.updateLinks(unlink(oddExisingChildren, keyPath, inverse: inverse))
-//            context.updateLinks(link(children, keyPath, inverse: inverse))
-//        }
-//    }
-}
-
-private extension EntityModelProtocol {
-
     func saveRelation<Child, Cardinality, Constraint>(
         at keyPath: KeyPath<Self, OneWayRelation<Child, Cardinality, Constraint>>,
         in context: inout Context) throws {
@@ -131,53 +100,4 @@ private extension EntityModelProtocol {
         try relation(keyPath).save(&context)
     }
 }
-//
-//private extension EntityModelProtocol {
-//    func link<Child, Cardinality, Constraint>(
-//        _ children: [Child.ID],
-//        _ keyPath: KeyPath<Self, OneWayRelation<Child, Cardinality, Constraint>>
-//
-//    ) -> Links<Self, Child> {
-//
-//        Links(
-//            direct: [Link(
-//                parent: id,
-//                children: children,
-//                attribute: LinkAttribute(
-//                    name: keyPath.name,
-//                    updateOption: relation(keyPath).directLinkUpdateOption
-//                )
-//            )],
-//            inverse: []
-//        )
-//    }
-//
-//    func link<Child, Cardinality, Constraint, InverseRelation, InverseConstraint>(
-//        _ children: [Child.ID],
-//        _ keyPath: KeyPath<Self, MutualRelation<Child, Cardinality, Constraint>>,
-//        inverse: KeyPath<Child, MutualRelation<Self, InverseRelation, InverseConstraint>>
-//
-//    ) -> Links<Self, Child> {
-//
-//        Links(
-//            direct: [Link(
-//                parent: id,
-//                children: children,
-//                attribute: LinkAttribute(
-//                    name: keyPath.name,
-//                    updateOption: relation(keyPath).directLinkUpdateOption
-//                )
-//            )],
-//            inverse: children.map { child in
-//                Link(
-//                    parent: child,
-//                    children: [id],
-//                    attribute: LinkAttribute(
-//                        name: inverse.name,
-//                        updateOption: MutualRelation<Self, InverseRelation, InverseConstraint>.inverseLinkUpdateOption
-//                    )
-//                )
-//            }
-//        )
-//    }
-//}
+
